+++ src/design/plone/contenttypes/tests/test_ct_documento.py
@@ -45,6 +45,7 @@ class TestDocument(unittest.TestCase):
                 "collective.dexteritytextindexer",
                 "plone.translatable",
                 "kitconcept.seo",
+                "plone.versioning",
             ),
         )